+++ src/Controller.js
@@ -1,6 +1,6 @@
-import Calculator from '../model/Calculator.js';
-import InputView from '../view/InputView.js';
-import OutputView from '../view/Outputview.js';
+import Calculator from './CalculatorModel.js';
+import InputView from './InputView.js';
+import OutputView from './OutputView.js';
 
 // InputView에서 입력을 받고 Model을 통해 처리한 후 OutputView로 출력
 // 컨트롤러가 상태(인스턴스 변수 등)를 관리하기 때문에 클래스로 정의
