+++ src/InputView.js
@@ -1,5 +1,5 @@
-import { INPUT_MESSAGE } from '../constant/message.js';
-import { readLineAsync } from '../util/consoleUtil.js';
+import { INPUT_MESSAGE } from './constant/message.js';
+import { readLineAsync } from './util/consoleUtil.js';
 
 // 사용자에게 입력을 받는 역할
 // 메서드 집합만을 제공하는 정적 객체
