+++ src/OutputView.js
@@ -1,4 +1,4 @@
-import { print } from '../util/consoleUtil.js';
+import { print } from './util/consoleUtil.js';
 
 // 결과를 사용자에게 출력하는 역할
 const OutputView = {
